Break out Supplies app from Project Settings folder.
