+++ supplies/views.py
