+++ supplies_tracker/urls.py
@@ -15,8 +15,8 @@ Including another URLconf
 """
 from django.conf.urls import url, include
 from django.contrib import admin
-from . import views
-from supplies_tracker import views as supplies_tracker_views
+from supplies import views
+from supplies import views as supplies_tracker_views
 from django.contrib.auth import views as auth_views
 # from django.urls import path
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
